% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ff_df_w_abstime.R
\name{build_ff_df_w_abstime}
\alias{build_ff_df_w_abstime}
\title{Build a dataframe out of a flowframe}
\usage{
build_ff_df_w_abstime(ff)
}
\arguments{
\item{ff}{The flowFrame}
}
\value{
a tidy dataframe of flow cytometry data; rows are observations, columns are channels; the identifier column is by default the name of the fcs file and is called 'exp'
}
\description{
This function takes a flowFrame object from the flowCore package as its only argument, and returns a tidy dataframe with the experiment data in columns with names according to the cytometer channels, a relative time column called 'Time', and an absolute time column in standard strptime format called 'abstime'.
}
