% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.analysis.R
\name{cc.analysis}
\alias{cc.analysis}
\title{pHluorin Calibration Curve Analysis}
\usage{
cc.analysis(subset, background, id, inputs = FALSE, buffer.values = FALSE,
  xmin = 4.9, xmax = 8.6, FITC.thresh = 800, BV.thresh = 800,
  start.list = list(a = 0.5, b = 2, c = 7, d = 0.25))
}
\arguments{
\item{subset}{The dataframe containing the raw calibration curve data. Identifier column should be called 'exp' and formatted as 'cc_pH.fcs'}

\item{background}{The dataframe containing the raw background data. Identifier column should be called 'exp' and formatted as 'experiment_strain_background' with background being either 'buffer' or 'media'}

\item{id}{Name/date of experiment, will end up as part of the cc function name so use plaintext characters and no dashes}

\item{inputs}{A list of column entries that were used during data collection, default is FALSE (inferred from filenames)}

\item{buffer.values}{A list of buffer values for that day, default is FALSE (inferred from filenames)}

\item{xmin}{minimum pH predicted (depends on range of buffer values used)}

\item{xmax}{maximum pH predicted (depends on range of buffer values used)}

\item{FITC.thresh}{Value that untransformed FITC.A (area) signal must exceed; default is 800}

\item{BV.thresh}{Value that untransformed BV510.A (area) signal must exceed; default is 800}

\item{start.list}{Edit the starting fitting parameters; default is list(a=.5, b=2, c=7, d=0.25) (keep this form but change numbers)}
}
\value{
convert.to.pH function, plot of calibration curve, analysis of fit quality
}
\description{
This function takes raw data from a calibration curve experiment (yeast expressing pHluorin, flow cytometry data) and does all processing steps to construct a calibration curve that maps fluorescence ratio to pH. Takes *raw* (untransformed) FITC and BV510 channels as inputs!
}
