% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.red.bkg.R
\name{generate.red.bkg}
\alias{generate.red.bkg}
\title{Get background red fluorescence for induction timecourses}
\usage{
generate.red.bkg(df, experiment.name = FALSE, media = "none")
}
\arguments{
\item{df}{Name of the dataframe which contains the induction data. Should be an output of read.chicago.flowSet or similar formatting.}

\item{experiment.name}{The precise name of the experiment to use to generate the background. Will be in the 'exp' column. Default value is FALSE, where the experiment name will be inferred.}

\item{media}{Is the background acidic or buffered neutral SC? If so, specify 'sc' or 'buffered.sc'; otherwise returns the median of all rows which have 'backgrounds' and 'ycgt028' in the 'exp' column.}
}
\value{
A single numerical value which is the median of the distribution of size-normalized red fluroescence per cell (PEDazzle area / forward scatter area).
}
\description{
From a dataframe representing a stress and recovery experiment (timecourse measuring induction of a red fluorophore-tagged protein), generate the background level of fluorescence in untreated cells.
}
