% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_rolling_mean.R
\name{pH_rolling_mean}
\alias{pH_rolling_mean}
\title{Calculate rolling mean of flow cytometry pH timecourse data}
\usage{
pH_rolling_mean(dat, k = "auto", k_val = 0.05)
}
\arguments{
\item{dat}{A dataframe or dataframe subset to act on. Should have a column called pH.m on which the fit will be calculated.}

\item{k}{Indicates method to use for binwidth. If k is an integer, this will be the number of samples to include in the averaging bin. If k is "auto", the binwidth will be determined for each subset individually, size will be determined by k_val}

\item{k_val}{Only used if k = "auto"; the percentage of the total number of observations that will be averaged over for each subset}
}
\value{
a dataframe with columns "Time" and "fit" representing the rolling mean at each timepoint.
}
\description{
Takes a subset of a dataframe that represents a single experiment and calculates a rolling mean on the data with binwidth k.
}
