% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.pH.induction.R
\name{process.pH.induction}
\alias{process.pH.induction}
\title{Processes a flow dataset of induction timeseries}
\usage{
process.pH.induction(df, cc = FALSE, summarise = FALSE,
  ref.strain = "ycgt028", ref.background = "media")
}
\arguments{
\item{df}{The dataframe containing the raw induction data. Identifier column should be called 'exp' and formatted as background data: 'controls_strain_background.fcs'; or induction data: 'timepoint_treatment_fitness_pH.fcs'}

\item{cc}{whether or not to generate a calibration curve and calculate intracellular pH; default is FALSE : CURRENTLY UNAVAILABLE}

\item{ref.strain}{the reference strain; default is "ycgt028"}

\item{ref.background}{the background of the reference strain; default is "media"}

\item{summarize}{function returns either a median-summarised timeseries (TRUE) or unsummarized data (FALSE; default)}

\item{channel}{induction channel name; default is "PEDazzle.A" : CURRENTLY UNAVAILABLE}
}
\value{
processed dataframe
}
\description{
This function takes a dataframe produced by merge_flowSet and produces a processed dataframe with induction data; currently requires pre-classification of populations into a column called 'population'
Remaining to do: 1. generate calibration curve and convert to pH within function; 2. find a way to change the induction channel name;
}
