% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.chicago.flowSet.R
\name{read.chicago.flowSet}
\alias{read.chicago.flowSet}
\title{Read in flow cytometry data generated at the UChicago Flow Cytometry Core}
\usage{
read.chicago.flowSet(filename, instrument = "nHTS", fullpath = FALSE)
}
\arguments{
\item{filename}{The name of the folder which contains the data. The function will default look in Cat Triandafillou's personal folder, so not useful for other users (sorry) but...}

\item{instrument}{Options are the HTS model ('HTS') or the other model ('nHTS'). 'nHTS' is default, specifying anything else will fail.}

\item{fullpath}{Should your data be located somewhere other than the default folder containing data, set this value to 'TRUE' and then put the full filepath instead of the folder name for the filename argument.}
}
\value{
tidy dataframe with all data from specified folder; individual FCS files appear in the 'exp' column, and all other data in the file appears in the dataframe.
}
\description{
Read a folder of FCS files generated on either the LSR Fortessa or the LSR Fortessa HTS at the flow cytometry core at the University of Chicago. Deals with the naming conventions of the individual instruments and produces a tidy dataframe with all specimen represented.
}
