% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fcs_cols.R
\name{rename_fcs_cols}
\alias{rename_fcs_cols}
\title{Rename the columns of a dataframe intelligently}
\usage{
rename_fcs_cols(dataframe, print.names = F)
}
\arguments{
\item{dataframe}{the name of the dataframe to be renamed}

\item{print.names}{whether or not to print the resulting column names; default is F}
}
\value{
a dataframe in the same format as merge_flowSet, but with shorter column names
}
\description{
Rename the columns of a dataframe intelligently
}
