% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ct_greypanels.R
\name{theme_ct_greypanels}
\alias{theme_ct_greypanels}
\title{Standard plotting theme}
\usage{
theme_ct_greypanels(base_size = 12, point.alpha = F)
}
\arguments{
\item{base_size}{text size; default is 12}

\item{point.alpha}{if T sets the alpha of the legend to be 1; default is F}
}
\description{
Adding this function to a ggplot changes the theme
}
