% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.analysis.R
\name{cc.analysis}
\alias{cc.analysis}
\title{pHluorin Calibration Curve Analysis}
\usage{
cc.analysis(subset, background, id, inputs, buffer.values, xmin = 4.9,
  xmax = 8.6)
}
\arguments{
\item{subset}{The dataframe containing the raw calibration curve data. Identifier column should be called 'exp' and formatted as 'cc_pH.fcs'}

\item{background}{The dataframe containing the raw background data. Identifier column should be called 'exp' and formatted as 'experiment_strain_background' with background being either 'buffer' or 'media'}

\item{id}{Name/date of experiment}

\item{inputs}{A list of column entries that were used during data collection}

\item{buffer.values}{A list of buffer values for that day}

\item{xmin}{minimum pH predicted (depends on range of buffer values used)}

\item{xmax}{maximum pH predicted (depends on range of buffer values used)}
}
\value{
convert.to.pH function, plot of calibration curve, analysis of fit quality
}
\description{
This function takes raw data from a calibration curve experiment (yeast expressing pHluorin, flow cytometry data) and does all processing steps to construct a calibration curve that maps fluorescence ratio to pH
}
