% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fcs_cols.R
\name{rename_fcs_cols}
\alias{rename_fcs_cols}
\title{Rename the columns of a dataframe intelligently}
\usage{
rename_fcs_cols(dataframe, print.names = F, method = "old")
}
\arguments{
\item{dataframe}{the name of the dataframe to be renamed}

\item{print.names}{whether or not to print the resulting column names; default is F}

\item{method}{method of column renaming; options are "old" (first iteration, better for non-HTS instrument, use with character swapped values), "new" (second version, better for HTS instrument, use with UNSWAPPED values); default is "old"}
}
\value{
a dataframe in the same format as merge_flowSet, but with shorter column names
}
\description{
Rename the columns of a dataframe intelligently
}
